Add redirection using an onClose callback in the hidden trap.
This was the original version I gave to students. The redirection had the intended effect of being so obtuse that it was hard for students to understand, but it was perhaps a bit too hard.

The version on the main branch is probably better for next time. It still requires students to use a conditional breakpoint, and even does a better job at preventing them from modifying anything in the variables pane than this version.